+++ R/ctrough.R
@@ -36,6 +36,7 @@ setClass(
 #' Last value.
 #'
 #' @inheritParams metrics_params
+#' @return an object of class last_metric
 #' @export
 Last <- function(variable = NULL, window = NULL, name = NULL, unit = NULL, stat_display = NULL, digits = NULL) {
   metric <- nca_constructor(
@@ -54,6 +55,7 @@ Last <- function(variable = NULL, window = NULL, name = NULL, unit = NULL, stat_
 #' Ctrough. Last time in x shall be considered as the trough time. Similar to Last, but for concentrations.
 #'
 #' @inheritParams metrics_params
+#' @return an object of class ctrough_metric
 #' @export
 Ctrough <- function(variable = NULL, window = NULL, name = NULL, unit = NULL, stat_display = NULL, digits = NULL) {
   metric <- nca_constructor(