+++ R/special_operators.R
@@ -1,5 +1,6 @@
 #' Magritt operator for piping.
 #'
+#' @return the result of the piping operation
 #' @importFrom magrittr `%>%`
 #' @export
 #' @keywords internal